03/03/2025 (2)

+++ src/main/java/com/minka/optica/services/patients/PatientsServiceImp.java
@@ -5,8 +5,14 @@ import com.minka.optica.dto.PatientsDto;
 import com.minka.optica.entities.Patients;
 import com.minka.optica.exceptions.BdNotFoundException;
 import com.minka.optica.exceptions.BdNotSaveException;
+import com.minka.optica.exceptions.DischargeDateException;
 import com.minka.optica.mapper.PatientsMapper;
 import com.minka.optica.repository.PatientsRepository;
+
+import java.time.LocalDate;
+import java.time.Period;
+import java.time.format.DateTimeFormatter;
+import java.time.temporal.ChronoUnit;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
@@ -80,8 +86,35 @@ public class PatientsServiceImp implements PatientsService{
   public Boolean deleteById(final Long id) {
     final Optional<Patients> existPatients = this.patientsRepository.findById(id);
     if (existPatients.isPresent()) {
-      this.patientsRepository.deleteById(id);
-      return true;
+      Patients patient = existPatients.get();
+
+      // Validar si dischargeDate es nulo o vacio
+      String dischargeDate = patient.getDischargeDate();
+      if (dischargeDate == null || dischargeDate.trim().isEmpty()) {
+        throw new DischargeDateException("The field dischargeDate is null or empty for the patient ID: " + id);
+      }
+
+      // Convertir String a LofechacalDate
+      DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
+      LocalDate fecha = LocalDate.parse(patient.getDischargeDate(), formatter);
+
+      // Fecha actual
+      LocalDate fechaActual = LocalDate.now();
+
+      // Calcular la diferencia en días
+      long diferenciaDias = ChronoUnit.DAYS.between(fecha, fechaActual);
+
+      // Resultado
+      System.out.println("Delete of patient. Difference in days: " + diferenciaDias);
+
+      // Calcular la diferencia
+      Period periodo = Period.between(fecha, fechaActual);
+
+      if (diferenciaDias > 1826) {
+        this.patientsRepository.deleteById(id);
+        return true;
+      }
+        return false;
     } else {
       throw new BdNotFoundException("DELETE - There is no patients in the database with the id: " + id);
     }
